+++ packages/react-querybuilder/src/components/Rule.tsx
@@ -497,6 +497,7 @@ export const useRule = (props: RuleProps): UseRule => {
     dragRef = null,
     isDragging = false,
     isOver = false,
+    dropNotAllowed = false,
   } = props;
 
   useDeprecatedProps('rule', !ruleProp);
@@ -746,6 +747,7 @@ export const useRule = (props: RuleProps): UseRule => {
         isOver && classNamesProp.dndOver,
         isOver && dropEffect === 'copy' && classNamesProp.dndCopy,
         isOver && groupItems && classNamesProp.dndGroup,
+        dropNotAllowed && classNamesProp.dndDropNotAllowed,
         hasSubQuery && classNamesProp.hasSubQuery,
         // standard conditional classes
         suppressStandardClassnames || {
@@ -754,6 +756,7 @@ export const useRule = (props: RuleProps): UseRule => {
           [standardClassnames.dndOver]: isOver,
           [standardClassnames.dndCopy]: isOver && dropEffect === 'copy',
           [standardClassnames.dndGroup]: isOver && groupItems,
+          [standardClassnames.dndDropNotAllowed]: dropNotAllowed,
           [standardClassnames.hasSubQuery]: hasSubQuery,
         },
         validationClassName
@@ -764,10 +767,12 @@ export const useRule = (props: RuleProps): UseRule => {
       classNamesProp.dndDragging,
       classNamesProp.dndGroup,
       classNamesProp.dndOver,
+      classNamesProp.dndDropNotAllowed,
       classNamesProp.hasSubQuery,
       classNamesProp.rule,
       disabled,
       dropEffect,
+      dropNotAllowed,
       fieldBasedClassName,
       fieldData,
       getRuleClassname,